+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.31.1] - 2026-07-21
+
 ### Changed
 
 - **Developer console (auto-open on startup)** — If a browser tab is already connected to the local viewer, KitLib no longer opens another one on startup.